+++ clients/native/src/commands/init.rs
@@ -12,6 +12,7 @@ use crypto::asymmetric::identity;
 use nymsphinx::addressing::clients::Recipient;
 use serde::Serialize;
 use std::fmt::Display;
+use std::net::IpAddr;
 use tap::TapFallible;
 
 #[derive(Args, Clone)]
@@ -30,7 +31,6 @@ pub(crate) struct Init {
     force_register_gateway: bool,
 
     /// Comma separated list of rest endpoints of the nyxd validators
-    #[cfg(feature = "coconut")]
     #[clap(long, alias = "nymd_validators", value_delimiter = ',')]
     nyxd_urls: Option<Vec<url::Url>>,
 
@@ -47,6 +47,10 @@ pub(crate) struct Init {
     #[clap(short, long)]
     port: Option<u16>,
 
+    /// Ip for the socket (if applicable) to listen for requests.
+    #[clap(long)]
+    host: Option<IpAddr>,
+
     /// Mostly debug-related option to increase default traffic rate so that you would not need to
     /// modify config post init
     #[clap(long, hide = true)]
@@ -58,7 +62,6 @@ pub(crate) struct Init {
 
     /// Set this client to work in a enabled credentials mode that would attempt to use gateway
     /// with bandwidth credential requirement.
-    #[cfg(feature = "coconut")]
     #[clap(long)]
     enabled_credentials_mode: Option<bool>,
 
@@ -73,12 +76,11 @@ impl From<Init> for OverrideConfig {
             nym_apis: init_config.nym_apis,
             disable_socket: init_config.disable_socket,
             port: init_config.port,
+            host: init_config.host,
             fastmode: init_config.fastmode,
             no_cover: init_config.no_cover,
 
-            #[cfg(feature = "coconut")]
             nyxd_urls: init_config.nyxd_urls,
-            #[cfg(feature = "coconut")]
             enabled_credentials_mode: init_config.enabled_credentials_mode,
         }
     }