+++ clients/native/src/commands/mod.rs
@@ -9,6 +9,7 @@ use completions::{fig_generate, ArgShell};
 use config::OptionalSet;
 use lazy_static::lazy_static;
 use std::error::Error;
+use std::net::IpAddr;
 
 pub(crate) mod init;
 pub(crate) mod run;
@@ -56,12 +57,10 @@ pub(crate) struct OverrideConfig {
     nym_apis: Option<Vec<url::Url>>,
     disable_socket: Option<bool>,
     port: Option<u16>,
+    host: Option<IpAddr>,
     fastmode: bool,
     no_cover: bool,
-
-    #[cfg(feature = "coconut")]
     nyxd_urls: Option<Vec<url::Url>>,
-    #[cfg(feature = "coconut")]
     enabled_credentials_mode: Option<bool>,
 }
 
@@ -78,35 +77,29 @@ pub(crate) async fn execute(args: &Cli) -> Result<(), Box<dyn Error + Send + Syn
     Ok(())
 }
 
-pub(crate) fn override_config(mut config: Config, args: OverrideConfig) -> Config {
-    config = config
+pub(crate) fn override_config(config: Config, args: OverrideConfig) -> Config {
+    config
         .with_optional(Config::with_disabled_socket, args.disable_socket)
         .with_base(BaseConfig::with_high_default_traffic_volume, args.fastmode)
         .with_base(BaseConfig::with_disabled_cover_traffic, args.no_cover)
         .with_optional(Config::with_port, args.port)
+        .with_optional(Config::with_host, args.host)
         .with_optional_custom_env_ext(
             BaseConfig::with_custom_nym_apis,
             args.nym_apis,
             network_defaults::var_names::NYM_API,
             config::parse_urls,
-        );
-
-    #[cfg(feature = "coconut")]
-    {
-        config = config
-            .with_optional_custom_env_ext(
-                BaseConfig::with_custom_nyxd,
-                args.nyxd_urls,
-                network_defaults::var_names::NYXD,
-                config::parse_urls,
-            )
-            .with_optional_ext(
-                BaseConfig::with_disabled_credentials,
-                args.enabled_credentials_mode.map(|b| !b),
-            );
-    }
-
-    config
+        )
+        .with_optional_custom_env_ext(
+            BaseConfig::with_custom_nyxd,
+            args.nyxd_urls,
+            network_defaults::var_names::NYXD,
+            config::parse_urls,
+        )
+        .with_optional_ext(
+            BaseConfig::with_disabled_credentials,
+            args.enabled_credentials_mode.map(|b| !b),
+        )
 }
 
 #[cfg(test)]