+++ common/cosmwasm-smart-contracts/vesting-contract/src/messages.rs
@@ -17,10 +17,7 @@ pub struct InitMsg {
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
 #[serde(rename_all = "snake_case")]
-pub struct MigrateMsg {
-    // I'm making it explicit so that we wouldn't accidentally forget about it
-    pub manually_verified_no_staking_addresses: bool,
-}
+pub struct MigrateMsg {}
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema, Default)]
 pub struct VestingSpecification {