+++ nym-api/src/support/storage/manager.rs
@@ -977,7 +977,6 @@ impl StorageManager {
     ///
     /// * `tx_hash`: hash of the deposit transaction.
     /// * `blinded_signature_response`: the encrypted blinded signature response.
-    #[cfg(feature = "coconut")]
     pub(crate) async fn insert_blinded_signature_response(
         &self,
         tx_hash: &str,
@@ -998,7 +997,6 @@ impl StorageManager {
     /// # Arguments
     ///
     /// * `tx_hash`: transaction hash of the deposit.
-    #[cfg(feature = "coconut")]
     pub(crate) async fn get_blinded_signature_response(
         &self,
         tx_hash: &str,