+++ nym-connect-android/src-tauri/src/operations/mod.rs
@@ -1,0 +1,7 @@
+pub mod connection;
+pub mod directory;
+pub mod export;
+pub mod help;
+pub mod http;
+#[cfg(desktop)]
+pub mod window;