+++ nym-connect/src-tauri/src/tasks.rs
@@ -9,7 +9,6 @@ use task::manager::TaskStatus;
 use tokio::sync::RwLock;
 
 use config_common::NymConfig;
-#[cfg(not(feature = "coconut"))]
 use nym_socks5::client::NymClient as Socks5NymClient;
 use nym_socks5::client::{config::Config as Socks5Config, Socks5ControlMessageSender};