+++ sdk/rust/nym-sdk/src/mixnet/client.rs
@@ -4,8 +4,8 @@ use client_connections::TransmissionLane;
 use client_core::{
     client::{
         base_client::{
-            non_wasm_helpers, BaseClientBuilder, ClientInput, ClientOutput, ClientState,
-            CredentialsToggle,
+            helpers::setup_empty_reply_surb_backend, non_wasm_helpers, BaseClientBuilder,
+            ClientInput, ClientOutput, ClientState, CredentialsToggle,
         },
         inbound_messages::InputMessage,
         key_manager::KeyManager,
@@ -348,10 +348,7 @@ impl MixnetClient {
         paths: Option<StoragePaths>,
     ) -> Result<MixnetClientBuilder<reply_storage::Empty>> {
         let config = config.unwrap_or_default();
-
-        let reply_storage_backend =
-            non_wasm_helpers::setup_empty_reply_surb_backend(&config.debug_config);
-
+        let reply_storage_backend = setup_empty_reply_surb_backend(&config.debug_config);
         MixnetClient::builder_with_custom_storage(Some(config), paths, reply_storage_backend)
     }