+++ sdk/typescript/packages/sdk/src/mixnet/wasm/index.ts
@@ -1,21 +1,23 @@
 import * as Comlink from 'comlink';
+import InlineWasmWebWorker from 'web-worker:./worker';
 import {
+  BinaryMessageReceivedEvent,
+  ConnectedEvent,
   EventHandlerFn,
   EventKinds,
   IWebWorker,
+  IWebWorkerAsync,
   IWebWorkerEvents,
-  ConnectedEvent,
   LoadedEvent,
-  StringMessageReceivedEvent,
-  BinaryMessageReceivedEvent,
   MimeTypes,
+  StringMessageReceivedEvent,
 } from './types';
 
 /**
  * Client for the Nym mixnet.
  */
 export interface NymMixnetClient {
-  client: Comlink.Remote<IWebWorker>;
+  client: IWebWorkerAsync;
   events: IWebWorkerEvents;
 }
 
@@ -24,7 +26,7 @@ export interface NymMixnetClient {
  *
  */
 export const createNymMixnetClient = async (options?: {
-  mimeTypes?: string[] | MimeTypes[];
+  autoConvertStringMimeTypes?: string[] | MimeTypes[];
 }): Promise<NymMixnetClient> => {
   // create a web worker that runs the WASM client on another thread and wait until it signals that it is ready
   // eslint-disable-next-line @typescript-eslint/no-use-before-define
@@ -90,11 +92,14 @@ export const createNymMixnetClient = async (options?: {
   };
 
   // let comlink handle interop with the web worker
-  const client = Comlink.wrap<IWebWorker>(worker);
+  const client: IWebWorkerAsync = Comlink.wrap<IWebWorker>(worker);
 
   // set any options
-  if (options?.mimeTypes) {
-    await client.setTextMimeTypes(options.mimeTypes);
+  if (options?.autoConvertStringMimeTypes) {
+    await client.setTextMimeTypes(options.autoConvertStringMimeTypes);
+  } else {
+    // set some sensible defaults for text mime types
+    await client.setTextMimeTypes([MimeTypes.ApplicationJson, MimeTypes.TextPlain]);
   }
 
   // pass the client interop and subscription manage back to the caller
@@ -112,9 +117,11 @@ export const createNymMixnetClient = async (options?: {
  */
 const createWorker = async () =>
   new Promise<Worker>((resolve, reject) => {
-    const worker = new Worker(
-      new URL('./worker.js', import.meta.url), // NB: this path is relative to the `dist` directory of this bundle
-    );
+    // rollup will inline the built worker script, so that when the SDK is used in
+    // other projects, they will not need to mess around trying to bundle it
+    // however, it will make this SDK bundle bigger because of Base64 inline data
+    const worker = new InlineWasmWebWorker();
+
     worker.addEventListener('error', reject);
     worker.addEventListener(
       'message',